refactor: useValidatedCoupons 훅 제거하고 순수 함수로 분리
