+++ src/utils/coupon.ts
@@ -1,13 +1,13 @@
 import { CartItemProps } from '../types/cartItem';
-import { Coupon } from '../types/coupon';
+import { Coupon, validatedCouponList } from '../types/coupon';
 
-export function isExpired(expirationDate: string): boolean {
+function isExpired(expirationDate: string): boolean {
   const today = new Date();
   const [y, m, d] = expirationDate.split('-').map(Number);
   return today <= new Date(y, m - 1, d);
 }
 
-export function isMiracleMorning(start: string, end: string): boolean {
+function isMiracleMorning(start: string, end: string): boolean {
   const now = new Date();
 
   const startDate = new Date(now);
@@ -23,15 +23,46 @@ export function isMiracleMorning(start: string, end: string): boolean {
   return now >= startDate && now <= endDate;
 }
 
-export function isMinimumAmount(
-  minimumAmount: number,
-  subTotal: number
-): boolean {
+function isMinimumAmount(minimumAmount: number, subTotal: number): boolean {
   return subTotal >= minimumAmount;
 }
 
-export function isQuantity(cartItems: CartItemProps[]): boolean {
+function isQuantity(cartItems: CartItemProps[]): boolean {
   return cartItems.some((item) => item.quantity >= 2);
+}
+
+export function validateCoupons(
+  couponList: Coupon[],
+  subTotal: number,
+  selectedItems: CartItemProps[]
+): validatedCouponList[] {
+  const validatedCouponList: validatedCouponList[] = [];
+
+  couponList.forEach((coupon) => {
+    let expiredFlag = !isExpired(coupon.expirationDate);
+
+    if (coupon.description.includes('미라클모닝')) {
+      expiredFlag = !isMiracleMorning(
+        coupon.availableTime?.start || '',
+        coupon.availableTime?.end || ''
+      );
+    }
+
+    if (
+      coupon.minimumAmount !== undefined &&
+      !isMinimumAmount(coupon.minimumAmount, subTotal)
+    ) {
+      expiredFlag = true;
+    }
+
+    if (coupon.buyQuantity !== undefined && coupon.getQuantity !== undefined) {
+      expiredFlag = !isQuantity(selectedItems);
+    }
+
+    validatedCouponList.push({ ...coupon, isExpired: expiredFlag });
+  });
+
+  return validatedCouponList;
 }
 
 function calFixedDiscount(discount: number): number {
